chore: refactor context exports
