+++ src/context/index.js
@@ -1,0 +1,2 @@
+export { useCategories } from "./categories-context";
+export { useVideos } from "./videos-context";
